// src/types/IUser.ts
export interface IUser {
    id: number;
    name: string;
    username: string;
    email: string;
    website: string;
    phone: string;
    address: IAddress;
    company: ICompany;
}

interface IAddress {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: IGeo;
}

interface IGeo {
    lat: string;
    lng: string;
}

interface ICompany {
    name: string;
    catchPhrase: string;
    bs: string;
}

export interface IUsers {
    results: IUser[];
}